Declare `PROC`, `FARPROC` as structs instead of delegates
These "delegates" from the metadata had a signature that had no bearing on the signature actually intended. Typing them as delegates in C# makes it *very* hard to convert them to delegates with another signature.
On the flipside, exposing them as structs with an `IntPtr` makes it very easy to either use `Marshal.GetDelegateForFunctionPointer` or C# native function pointers to invoke the function at that address.

Fixes #396