Remove `HRESULT_FROM_WIN32` from HRESULT type
As a member of the HRESULT struct, it caused the otherwise small HRESULT struct to also generate the very large `WIN32_ERROR` enum, even if the user wouldn't have used this function.
So instead, we start a new concept where 'macros' such as this can be defined as a template, and then requested for generation explicitly. The generator is smart enough to detect what types the macro uses and generate them as well, as proven by the new test.