Fix lang service reporting C# 7.3 errors in Error List
By setting LangVersion so late in msbuild evaluation, the `CommandLineArgsForDesignTimeEvaluation` property is set based on the default value (of 7.3, for .NET Framework projects). This causes the VS language service to initialize with the wrong language version.

I originally moved this property definition to the .targets because `$(Language)` isn't set in time for the .props files to use them. But we can move it back to .props by using the project file extension instead.